reorg(r9c): break the platform/supervisor cycle, relocate assets (#489)
core/platform/service_manager.py and core/llm/providers/ollama.py
imported each other across the tier boundary — seven edges, every one a
function-local import deferred purely to dodge the cycle.

ollama.py was misfiled. Its own docstring calls it a "Host-native Ollama
supervisor"; it implements service_manager's ServiceSpec/ServiceStatus/
ActionResult protocol, and service_manager was its only consumer —
nothing in core/llm/ imported it. It becomes
core/platform/ollama_supervisor.py.

That alone leaves a manager <-> supervisor cycle, so the shared types
move to core/platform/service_contract.py, mirroring the split
core/federation/contract.py already uses for the same reason.
service_manager re-exports them, so existing callers are unaffected. It
now has zero deferred first-party imports; the supervisor reaches back
for SERVICES only inside its CLI main().

The move also fixes a bug the reorg introduced. REPO_ROOT was
parents[1], correct when the file was services/ollama_process.py. At
core/llm/providers/ollama.py that resolved to <repo>/core/llm, so the
pidfile and logfile were written to core/llm/logs/ — silently, because
.gitignore matches a bare logs/ at any depth, leaving _managed_by_vigil()
reading a pidfile nothing writes. Now parents[2], i.e. <repo>/logs.

Also:
- core/auth/extension_* -> core/integrations/extension/ as Connector
  Trust. It shares no symbol with the user-identity stack r9a put in
  core/auth/ and reaches for integration secrets instead, so its one
  cross-package dependency becomes an intra-package one.
- tools/security_detections.py -> core/detections/tools.py. The
  mechanism is tool provision, but the knowledge is detections.
- workflows/ -> core/workflows/definitions/, so the playbooks ship with
  the domain that reads them.

data/ stays at the repo root, against the diagram. It is not a static
asset directory: it holds the orchestrator workdirs
(data/investigations), the MCP tool cache, and the JSON storage
fallback. Moving it under core/ would put runtime writes inside a Python
package. Only its six bundled files fit the diagram's intent, and
splitting them out to leave a second data/ reads worse than leaving it.

Unit+llm 1211 passed / 4 failed — the same four pre-existing failures.
No new F401/F821.
