+++ core/detections/tools.py
