+++ core/integrations/extension/__init__.py
@@ -1,0 +1,2 @@
+"""Connector Trust — which page-extension connector origins Vigil admits
+into its own page, and the short-lived tokens minted to them."""
