+++ core/workflows/definitions/forensic-analysis/WORKFLOW.md
