+++ core/workflows/definitions/full-investigation/WORKFLOW.md
