+++ services/api/routers/extensions.py
@@ -12,7 +12,7 @@ from fastapi import APIRouter, Depends, HTTPException
 from services.api.middleware.auth import get_current_active_user
 from core.auth.auth_service import AuthService
 from core.storage.models import User
-from core.auth import extension_session_service as ext_session
+from core.integrations.extension import session_service as ext_session
 from core.routing import Auth, RouterMeta
 
 logger = logging.getLogger(__name__)
